feat: add ceramic-tests binary to be a test driver
This change adds a Rust binary `ceramic-tests` that is the driver for tests.
It handles creating the network and running the test job.

This driver also now support running a specific test image so we can
ensure that we are testing the current code state as opposed to the
latest published state.

The CI jobs have been updated to use this new driver.