feat: add smoke test suite and update hermetic driver
This change represents a bit of a paradigm shift in the test strategy.
Previously we had intended to have a single test driver and multiple
test suites in order to cover the matrix of needed tests.

Now the design is to have multiple tests drivers and a single test
suite. This way its clear that all tests (with rare exceptions) run
against both kind of infrastructure. The tests themselves are decoupled
from the kind of infrastructure they target.

The two kinds of infrastructure are:

Hermetic - isolated infrastructure
Durable - real world public infrastructure

See the README for more details.