feat: auto update deps as part of test run
This change introduces a script that updates the 3box deps of the test
suite so that we are always testing against latest released code.

Updating the deps initially made all the tests fail. The fix was to
remove the `moduleFileNames` config from the `jest.config.json` file, as
it was placing `.ts` files first causing deps that do not compile the
`.ts` files as ESM modules to fail vs just using the `.js` module files
in the deps.